Refactor: Extract create-image action (#57)
This extracts parts of the logic introduced via #39 to set up and deploy a one-click image into a new GitHub action I have created right now for just this purpose, LinqLover/create-image. In addition, an infinite loop in the prepare script is resolved that could occur if the repository was not found (ce05325).